Add loading effect to ProductCard components
Also, breaks the card into to new components: CardInfo (that display the
price and name of the product) and ImageBox (that displays an image).
Both components has the loading effect.
ImageBox can also be used as profile picture since it does not have a
fixed sizing.

This commit is part of issue #4